refactor: reorganize orchestrator package and eliminate legacy code
- Split orchestrator.go (186 lines) into 5 focused files:
  - coordinator.go (110 lines): Core orchestration logic
  - options.go (52 lines): Functional options pattern
  - types.go (17 lines): Type aliases
  - reconcile.go (32 lines): Domain reconciliation
  - apply_legacy.go: Renamed from apply.go

- Eliminated legacy code by moving domain logic:
  - Created packages/apply.go: Package installation logic
  - Created dotfiles/apply.go: Dotfile management logic
  - Removed apply_legacy.go entirely

- Achieved clean architecture:
  - Orchestrator now purely coordinates
  - Domain logic lives in domain packages
  - No functional changes, all tests pass
