+++ internal/orchestrator/coordinator.go
@@ -9,8 +9,6 @@ import (
 
 	"github.com/richhaase/plonk/internal/config"
 	"github.com/richhaase/plonk/internal/lock"
-	"github.com/richhaase/plonk/internal/output"
-	"github.com/richhaase/plonk/internal/resources"
 	"github.com/richhaase/plonk/internal/resources/dotfiles"
 	"github.com/richhaase/plonk/internal/resources/packages"
 )
@@ -27,55 +25,6 @@ type Orchestrator struct {
 	packagesOnly bool
 	dotfilesOnly bool
 }
-
-// Option is a functional option for configuring the orchestrator
-type Option func(*Orchestrator)
-
-// WithConfig sets the configuration
-func WithConfig(cfg *config.Config) Option {
-	return func(o *Orchestrator) {
-		o.config = cfg
-	}
-}
-
-// WithConfigDir sets the config directory
-func WithConfigDir(dir string) Option {
-	return func(o *Orchestrator) {
-		o.configDir = dir
-	}
-}
-
-// WithHomeDir sets the home directory
-func WithHomeDir(dir string) Option {
-	return func(o *Orchestrator) {
-		o.homeDir = dir
-	}
-}
-
-// WithDryRun enables dry-run mode
-func WithDryRun(dryRun bool) Option {
-	return func(o *Orchestrator) {
-		o.dryRun = dryRun
-	}
-}
-
-// WithPackagesOnly applies packages only
-func WithPackagesOnly(packagesOnly bool) Option {
-	return func(o *Orchestrator) {
-		o.packagesOnly = packagesOnly
-	}
-}
-
-// WithDotfilesOnly applies dotfiles only
-func WithDotfilesOnly(dotfilesOnly bool) Option {
-	return func(o *Orchestrator) {
-		o.dotfilesOnly = dotfilesOnly
-	}
-}
-
-// ApplyResult represents the result of an apply operation
-// ApplyResult is now defined in output package
-type ApplyResult = output.ApplyResult
 
 // New creates a new orchestrator instance with options
 func New(opts ...Option) *Orchestrator {
@@ -114,7 +63,7 @@ func (o *Orchestrator) Apply(ctx context.Context) (ApplyResult, error) {
 
 	// Apply packages (unless dotfiles-only)
 	if !o.dotfilesOnly {
-		packageResult, err := ApplyPackages(ctx, o.configDir, o.config, o.dryRun)
+		packageResult, err := packages.Apply(ctx, o.configDir, o.config, o.dryRun)
 		result.Packages = &packageResult
 		if err != nil {
 			result.AddPackageError(fmt.Errorf("package apply failed: %w", err))
@@ -123,7 +72,7 @@ func (o *Orchestrator) Apply(ctx context.Context) (ApplyResult, error) {
 
 	// Apply dotfiles (unless packages-only)
 	if !o.packagesOnly {
-		dotfileResult, err := ApplyDotfiles(ctx, o.configDir, o.homeDir, o.config, o.dryRun)
+		dotfileResult, err := dotfiles.Apply(ctx, o.configDir, o.homeDir, o.config, o.dryRun)
 		result.Dotfiles = &dotfileResult
 		if err != nil {
 			result.AddDotfileError(fmt.Errorf("dotfile apply failed: %w", err))
@@ -161,25 +110,4 @@ func (o *Orchestrator) Apply(ctx context.Context) (ApplyResult, error) {
 	}
 
 	return result, nil
-}
-
-// ReconcileAll reconciles all domains - coordination logic belongs in orchestrator
-func ReconcileAll(ctx context.Context, homeDir, configDir string) (map[string]resources.Result, error) {
-	results := make(map[string]resources.Result)
-
-	// Reconcile dotfiles using domain package
-	dotfileResult, err := dotfiles.Reconcile(ctx, homeDir, configDir)
-	if err != nil {
-		return nil, err
-	}
-	results["dotfile"] = dotfileResult
-
-	// Reconcile packages using domain package
-	packageResult, err := packages.Reconcile(ctx, configDir)
-	if err != nil {
-		return nil, err
-	}
-	results["package"] = packageResult
-
-	return results, nil
 }
