+++ internal/orchestrator/reconcile.go
@@ -1,0 +1,33 @@
+// Copyright (c) 2025 Rich Haase
+// Licensed under the MIT License. See LICENSE file in the project root for license information.
+
+package orchestrator
+
+import (
+	"context"
+
+	"github.com/richhaase/plonk/internal/resources"
+	"github.com/richhaase/plonk/internal/resources/dotfiles"
+	"github.com/richhaase/plonk/internal/resources/packages"
+)
+
+// ReconcileAll reconciles all domains - coordination logic belongs in orchestrator
+func ReconcileAll(ctx context.Context, homeDir, configDir string) (map[string]resources.Result, error) {
+	results := make(map[string]resources.Result)
+
+	// Reconcile dotfiles using domain package
+	dotfileResult, err := dotfiles.Reconcile(ctx, homeDir, configDir)
+	if err != nil {
+		return nil, err
+	}
+	results["dotfile"] = dotfileResult
+
+	// Reconcile packages using domain package
+	packageResult, err := packages.Reconcile(ctx, configDir)
+	if err != nil {
+		return nil, err
+	}
+	results["package"] = packageResult
+
+	return results, nil
+}
